+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added method `untype()` to return an untyped collection
+
 ## 1.0
 
 Initial release